refactor : memberService와 kakaoService 관련 로직 분리
- 기존의 memberService의 kakao 관련 로직 kakaoService로 분할

+++ playwithme/src/main/java/com/idea5/playwithme/member/domain/Member.java
@@ -40,7 +40,7 @@ public class Member {
     @CreatedDate
     private LocalDateTime createdAt;
 
-    private Integer mannerTemp;
+    private Float mannerTemp;
 
     @Column(length = 20)
     private String gender;
